6 Organize src directory
