--- src/main/scala/Compiler.scala
@@ -1,10 +1,0 @@
-package scruby
-
-object Compiler {
-  def apply(code: String): Either[CompilationError, List[SyntaxTree]] = {
-    for {
-      tokens <- Lexer(code).right
-      trees <- ScrubyParser(tokens).right
-    } yield trees
-  }
-}
