+++ src/main/scala/com/ethowitz/scruby/core/ScrubyFalseClass.scala
@@ -1,4 +1,4 @@
-package scruby
+package com.ethowitz.scruby.core
 
 object ScrubyFalseClass extends ScrubyObject('FalseClass, None, Map()) {
   override def toString: String = "false"
