+++ src/main/scala/com/ethowitz/scruby/core/ScrubyNilClass.scala
@@ -1,0 +1,11 @@
+package com.ethowitz.scruby.core
+
+import com.ethowitz.scruby.evaluator.ScrubyMethod
+import com.ethowitz.scruby.parser.ScrubyObjectContainer
+
+object ScrubyNilClass extends ScrubyObject('NilClass, None, Map()) {
+  override def toString: String = "nil"
+
+  override def predefMethods: Map[Symbol, ScrubyMethod] =
+    super.predefMethods + (Symbol("nil?") -> ScrubyMethod(ScrubyObjectContainer(ScrubyTrueClass)))
+}
