+++ src/main/scala/com/ethowitz/scruby/core/ScrubySymbol.scala
@@ -1,4 +1,6 @@
-package scruby
+package com.ethowitz.scruby.core
+
+import com.ethowitz.scruby.evaluator.ScrubyMethod
 
 class ScrubySymbol(val s: Symbol, ms: Map[Symbol, ScrubyMethod])
     extends ScrubyObject('Class, Some('Symbol), ms) {
