+++ src/main/scala/com/ethowitz/scruby/core/ScrubyTrueClass.scala
@@ -1,4 +1,4 @@
-package scruby
+package com.ethowitz.scruby.core
 
 object ScrubyTrueClass extends ScrubyObject('TrueClass, None, Map()) {
   override def toString: String = "true"
