+++ src/main/scala/com/ethowitz/scruby/exceptions/CompilationError.scala
@@ -1,0 +1,3 @@
+package com.ethowitz.scruby.exceptions
+
+trait CompilationError
