+++ src/main/scala/com/ethowitz/scruby/exceptions/LexerError.scala
@@ -1,0 +1,3 @@
+package com.ethowitz.scruby.exceptions
+
+case class LexerError(msg: String) extends CompilationError
