+++ src/main/scala/com/ethowitz/scruby/exceptions/ParserError.scala
@@ -1,0 +1,3 @@
+package com.ethowitz.scruby.exceptions
+
+case class ParserError(msg: String) extends CompilationError
