+++ src/main/scala/com/ethowitz/scruby/lexer/Token.scala
@@ -1,4 +1,4 @@
-package scruby
+package com.ethowitz.scruby.lexer
 
 sealed trait Token
